+++ docs/Changelog.md
@@ -5,7 +5,8 @@ This page is a brief overview of each version.
 ---
 ## Unreleased
 - Don't index and generate embeddings for pages in Library/
-- Add new `AI: Enhance Note` command to call existing `AI: Tag Note` and `AI: Suggest Page Name` commands on a note
+- Add new `AI: Enhance Note` command to call existing `AI: Tag Note` and `AI: Suggest Page Name` commands on a note, and the new frontmatter command
+- Add new `AI: Generate Note FrontMatter` command to extract useful facts from a note and add them to the frontmatter
 
 ---
 ## 0.2.0